# backend/booking-orchestrator/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Booking Orchestrator API",
    "description": "Main orchestration service that coordinates authentication, bookings, and course management",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/courses": {
      "get": {
        "operationId": "getAllCourses",
        "tags": ["courses"],
        "summary": "Get all courses",
        "description": "Returns a list of all available courses.",
        "responses": {
          "200": {
            "description": "Courses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Course"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/courses/{id}": {
      "get": {
        "operationId": "getCourseById",
        "tags": ["courses"],
        "summary": "Get course by ID",
        "description": "Returns detailed information about a specific course.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "course": {
                          "$ref": "#/components/schemas/Course"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/bookings": {
      "get": {
        "operationId": "getUserBookings",
        "tags": ["bookings"],
        "summary": "Get user bookings",
        "description": "Returns all bookings for the authenticated user.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bookings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "bookings": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Booking"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/bookings/start": {
      "post": {
        "operationId": "startBookingProcess",
        "tags": ["bookings"],
        "summary": "Start booking process",
        "description": "Initiates the booking process by creating a reservation and payment order.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "courseId": {
                    "type": "string"
                  },
                  "slotIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": ["courseId", "slotIds"]
              },
              "examples": {
                "startBookingRequest": {
                  "summary": "Example start booking request",
                  "value": {
                    "courseId": "695ba111b2e63b75bee66279",
                    "slotIds": ["695ba111b2e63b75bee6627b"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking process started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "reservationId": {
                          "type": "string"
                        },
                        "priceToPay": {
                          "type": "string"
                        },
                        "orderID": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "startBookingResponse": {
                    "summary": "Example start booking response",
                    "value": {
                      "data": {
                        "reservationId": "6969808214d2f13caa1e5f16",
                        "priceToPay": "49.00",
                        "orderID": "8VX12345AB6789012"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/bookings/confirm": {
      "post": {
        "operationId": "confirmBooking",
        "tags": ["bookings"],
        "summary": "Confirm booking",
        "description": "Confirms a booking by capturing payment and creating the final booking record. If booking fails after payment capture, automatically initiates a refund.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderID": {
                    "type": "string"
                  },
                  "reservationId": {
                    "type": "string"
                  }
                },
                "required": ["orderID", "reservationId"]
              },
              "examples": {
                "confirmBookingRequest": {
                  "summary": "Example confirm booking request",
                  "value": {
                    "orderID": "8VX12345AB6789012",
                    "reservationId": "6969808214d2f13caa1e5f16"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "booking": {
                          "$ref": "#/components/schemas/Booking"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/bookings/reservations/pending": {
      "post": {
        "operationId": "getPendingReservation",
        "tags": ["bookings"],
        "summary": "Get pending reservation",
        "description": "Retrieves the active pending reservation for the authenticated user and specified course.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "courseId": {
                    "type": "string"
                  }
                },
                "required": ["courseId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pending reservation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Reservation"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/bookings/reservations/{reservationId}/cancel": {
      "patch": {
        "operationId": "cancelReservation",
        "tags": ["bookings"],
        "summary": "Cancel reservation",
        "description": "Cancels a pending reservation for the authenticated user.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "reservationId",
            "required": true,
            "description": "Reservation ID to cancel",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The server could not understand the request due to invalid syntax",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required and has failed or not been provided",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "The client does not have access rights to the content",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "The server cannot find the requested resource",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "The request could not be completed due to a conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "The server encountered an unexpected condition",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "BadGateway": {
        "description": "The server received an invalid response from the upstream server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The server is currently unavailable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "The server did not receive a timely response from the upstream server",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NoContent": {
        "description": "The request was successful, but there is no content to return"
      }
    },
    "schemas": {
      "Course": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "slots": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "start": {
                  "type": "string",
                  "format": "date-time"
                },
                "end": {
                  "type": "string",
                  "format": "date-time"
                },
                "capacity": {
                  "type": "integer"
                },
                "available": {
                  "type": "integer"
                }
              }
            }
          },
          "open": {
            "type": "boolean"
          },
          "priceOptions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "numberSlots": {
                  "type": "integer"
                },
                "price": {
                  "type": "string"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "course": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "reservation": {
            "type": "string"
          },
          "slots": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": ["confirmed", "cancelled"]
          },
          "transactionId": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "course": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "slots": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": ["held", "completed", "cancelled", "expired"]
          },
          "expiration": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
...
}